+++ CHANGELOG.md
@@ -7,8 +7,9 @@ None
 ## [WIP] 0.3.0 / 2019-January-??
 
 - Add inf-clojure (Inferior Clojure mode)
-- Enable inf-clojure by default
-- Disable CIDER by default
+  - Enable inf-clojure by default
+  - Disable CIDER by default
+- Add Vue.js mode
 
 
 ## 0.2.1 / 2017-December-31